feat: 🎸 ts->js
