--- src/i18n.ts
@@ -1,24 +1,0 @@
-import i18n from "i18next";
-import { initReactI18next } from "react-i18next";
-
-import en from "./locales/en.json";
-import ja from "./locales/ja.json";
-
-const resources = {
-  en: {
-    translation: en,
-  },
-  ja: {
-    translation: ja,
-  },
-};
-
-i18n.use(initReactI18next).init({
-  resources,
-  fallbackLng: "en",
-  interpolation: {
-    escapeValue: false,
-  },
-});
-
-export default i18n;
